+++ src/main/java/project_pet_backEnd/groomer/controller/GroomerController.java
@@ -1,4 +1,11 @@
 package project_pet_backEnd.groomer.controller;
 
+
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@RequestMapping("/manager")
 public class GroomerController {
+
+
+
 }
