+++ src/main/java/project_pet_backEnd/groomer/dto/GetAllGroomerResponse.java
@@ -1,0 +1,11 @@
+package project_pet_backEnd.groomer.dto;
+
+import lombok.Data;
+
+@Data
+public class GetAllGroomerResponse {
+    private String pgName;
+    private String pgPic;
+
+    // 此處省略建構子、Getter 和 Setter 方法
+}
