+++ src/main/java/project_pet_backEnd/groomer/dto/PetGroomerInsertRequest.java
@@ -1,12 +1,11 @@
 package project_pet_backEnd.groomer.dto;
 
-import lombok.AllArgsConstructor;
+
 import lombok.Data;
 
 import java.sql.Date;
 
 @Data
-@AllArgsConstructor
 public class PetGroomerInsertRequest {
     private int manId;
     private String pgName;
